+++ keyper/const.py
@@ -1,0 +1,3 @@
+_PASSWORD_ALPHABET = 'abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789!@£$%^&*()_+-={}[]:|;<>?,./~`'
+
+__version__ = '0.3'
